refactored ast optimizations into optimizer mod
