+++ bin/scribe/src/main.rs
@@ -1,5 +1,5 @@
-use papyrus::ast_optimization::optimize_ast;
 use papyrus::miden_generator;
+use papyrus::optimizer::optimize_ast;
 use papyrus::parser;
 use papyrus::type_inference::infer_types;
