+++ crates/papyrus/src/lib.rs
@@ -1,6 +1,6 @@
-pub mod ast_optimization;
 pub mod executor;
 pub mod miden_generator;
+pub mod optimizer;
 pub mod parser;
 pub mod type_inference;
 pub mod types;
